+++ lib/less/rails.rb
@@ -1,6 +1,5 @@
 module Less
   module Rails
-    autoload :LessTemplate, 'less/rails/template_handlers'
   end
 end
 
@@ -12,4 +11,6 @@ require 'sprockets/railtie'
 
 require 'less/rails/version'
 require 'less/rails/helpers'
+require 'less/rails/template_handlers'
+require 'less/rails/css_compressor'
 require 'less/rails/railtie'